+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Next release
 (Add highlights/major features below)
 
+- server: log locking DB queries during source catalog migration
+
 ### Bug fixes and improvements
 (Add entries below in the order of server, console, cli, docs, others)